refactor: Update TeacherTestForm to include courseId prop
This refactor updates the TeacherTestForm component in the TeacherTestForm.tsx file to include a courseId prop. The prop is passed down from the TestFormPage component and is used to set the course value in the form. This enhancement ensures that the correct course is associated with the test being created or edited.